Updated ColorRoles-command.action
I added a few additional subcommands. I was not able to bring all of my ideas into fruition and so I will slowly work on them over time and may submit an addon snippet or attempt to resubmit the completed action later on. this will probably take me a few months.